+++ tests/testthat/test-sts_join.R
@@ -1,4 +1,13 @@
-test_that("joining sts", {
+
+test_that("input is validated", {
+
+  expect_error(
+    sts_join()
+  )
+
+  expect_error(
+    sts_join(example_sts, list("not", "an", "mts"))
+  )
 
   raws_sts <- example_raws
 
@@ -12,18 +21,18 @@ test_that("joining sts", {
     sts_join(raws_sts, example_sts)
   )
 
-  # Join with one object not being a sts
-  expect_error(
-    sts_join(example_sts, list("not", "a", "sts"))
-  )
+})
 
-  # Joining a sts with itself should be the original sts
+
+test_that("simple joining works", {
+
+  # Joining an sts with itself should be the original sts
   expect_identical(
     sts_join(example_sts, example_sts) %>% sts_extractData(),
     example_sts %>% sts_extractData()
   )
 
-  # Joining two partitions of a sts
+  # Join 2
   sts_1 <- sts_filterDate(example_sts, 20180801, 20180814)
   sts_2 <- sts_filterDate(example_sts, 20180814, 20180828)
 
@@ -32,7 +41,7 @@ test_that("joining sts", {
     example_sts %>% sts_extractData()
   )
 
-  # Joining three partitionsof a sts
+  # Join 3
   sts_1 <- sts_filterDate(example_sts, 20180801, 20180810)
   sts_2 <- sts_filterDate(example_sts, 20180810, 20180820)
   sts_3 <- sts_filterDate(example_sts, 20180820, 20180828)